report: move marshal() test helper to helpers_test.go
Relocated marshal() from application_test.go to helpers_test.go
so it can be reused by multiple test files.
